Revert "от AdService отделен CommentService"
