+++ src/main/java/ru/skypro/homework/controller/UserController.java
@@ -45,6 +45,11 @@ public class UserController {
                             responseCode = "403",
                             description = "Доступ запрещен",
                             content = @Content()
+                    ),
+                    @ApiResponse(
+                            responseCode = "404",
+                            description = "Пользователь не найден",
+                            content = @Content()
                     )
             }
     )
