+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -26,6 +26,16 @@ public class AdServiceImpl implements AdService {
     }
 
     @Override
+    public CommentsDto getComments(Integer id) {
+        return null;
+    }
+
+    @Override
+    public CreateOrUpdateCommentDto addComment(Integer id, CreateOrUpdateCommentDto newComment) {
+        return null;
+    }
+
+    @Override
     public ExtendedAdDto getAds(Integer id) {
         return null;
     }
@@ -41,6 +51,16 @@ public class AdServiceImpl implements AdService {
     }
 
     @Override
+    public void deleteComment(Integer adId, Integer commentId) {
+
+    }
+
+    @Override
+    public CreateOrUpdateCommentDto updateComment(Integer adId, Integer commentId, CreateOrUpdateCommentDto comment) {
+        return null;
+    }
+
+    @Override
     public AdsDto getMyAds() {
         return null;
     }
@@ -48,5 +68,8 @@ public class AdServiceImpl implements AdService {
     @Override
     public void updateImage(Integer id, MultipartFile image) {
 
+    }
+    public Comment getByAdId(Integer id) {
+        return commentRepository.findByAdId(id);
     }
 }
